Move images into src/images and update image imports
