+++ src/fixedMenu.js
@@ -7,7 +7,7 @@ import {
 } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
 import { BetaWarning } from './Beta';
-import logo from './logo.svg';
+import logo from './images/logo.svg';
 
 const FixedMenu = () => (
   <Menu fixed='top' size='large' color='teal' inverted >
