+++ src/loginPage.js
@@ -24,7 +24,7 @@ const LoginPage = () => (
     >
       <Grid.Column style={{ maxWidth: 450 }}>
         <Header as='h2' color='teal' textAlign='center'>
-          <Image src='/logo.svg' />
+          <Image src='./images/logo.svg' />
           {' '}Log-in to your account
         </Header>
         <Form size='large'>
