+++ src/simpleMenu.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Menu, Image, Header } from 'semantic-ui-react';
 import { Link } from 'react-router-dom';
 import { BetaWarning } from './Beta';
-import logo from './logo.svg';
+import logo from './images/logo.svg';
 
 // const SimpleMenu = (props) => (
